+++ Zephyrus/src/main/java/net/lordsofcode/zephyrus/spells/LifeSteal.java
@@ -49,7 +49,7 @@ public class LifeSteal extends Spell {
 	@Override
 	public void run(Player player, String[] args) {
 		Creature en = (Creature) getTarget(player);
-		en.damage((double)2);
+		en.damage(2);
 		try {
 			player.setHealth(player.getMaxHealth() + 2);
 		} catch (Exception e) {
